server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
cloudinary
python-multipart